+++ lib/features/movie_detail/domain/entities/gallery/image_info_entity.dart
@@ -1,5 +1,5 @@
 class ImageInfoEntity {
-  final int aspectRatio;
+  final double aspectRatio;
   final String filePath;
   final int height;
   final int width;
